+++ src/providers/ExtensionDetectionProvider/detectors/domElementDetector.ts
@@ -1,0 +1,6 @@
+import type { ExtensionDetector } from '../types';
+
+export const domElementDetector = (selector: string): ExtensionDetector => ({
+  strategy: `dom-element:${selector}`,
+  detect: async () => document.querySelector(selector) !== null,
+});
