+++ src/providers/ExtensionDetectionProvider/detectors/resourceFetchDetector.ts
@@ -1,0 +1,15 @@
+import type { ExtensionDetector } from '../types';
+
+export const resourceFetchDetector = (
+  resourceUrl: string,
+): ExtensionDetector => ({
+  strategy: `resource-fetch:${resourceUrl}`,
+  detect: async () => {
+    try {
+      const res = await fetch(resourceUrl, { method: 'HEAD' });
+      return res.ok;
+    } catch {
+      return false;
+    }
+  },
+});
