+++ src/providers/ExtensionDetectionProvider/extensionDetectionInitialDefinitions.ts
@@ -1,13 +1,9 @@
-import type { ExtensionDefinition } from './utils';
+import { createExtensionDetector } from './createExtensionDetector';
 import {
-  chromeExtensionInjectedDetector,
-  eip6963AnnounceProviderDetector,
   POCKET_UNIVERSE_EIP6963_LISTEN_MS,
   POCKET_UNIVERSE_HTML_DATA_CSN_SNAPSHOT_OBSERVE_MS,
-  pocketUniverseDatasetCsnDetector,
-  pocketUniverseHtmlDataCsnSnapshotDetector,
-  postMessageProxyDetector,
-} from './utils';
+} from './detectors/pocketUniverse/htmlDataCsnDetector';
+import type { ExtensionDefinition } from './types';
 
 export const POCKET_UNIVERSE_EXTENSION = 'pocket';
 
@@ -16,16 +12,21 @@ const POCKET_UNIVERSE_EXTENSION_ID = 'gacgndbocaddlemdiaadajmlggabdeod';
 export const pocketUniverseExtensionDefinition: ExtensionDefinition = {
   name: POCKET_UNIVERSE_EXTENSION,
   detectors: [
-    pocketUniverseHtmlDataCsnSnapshotDetector(),
-    postMessageProxyDetector(),
-    chromeExtensionInjectedDetector(POCKET_UNIVERSE_EXTENSION_ID),
-    eip6963AnnounceProviderDetector(
-      { nameIncludes: 'Pocket Universe' },
-      POCKET_UNIVERSE_EIP6963_LISTEN_MS,
-    ),
-    pocketUniverseDatasetCsnDetector(
-      POCKET_UNIVERSE_HTML_DATA_CSN_SNAPSHOT_OBSERVE_MS,
-    ),
+    createExtensionDetector({ kind: 'pocketHtmlDataCsnSnapshot' }),
+    createExtensionDetector({ kind: 'pocketPostMessageProxy' }),
+    createExtensionDetector({
+      kind: 'chromeExtensionInjected',
+      extensionId: POCKET_UNIVERSE_EXTENSION_ID,
+    }),
+    createExtensionDetector({
+      kind: 'eip6963AnnounceProvider',
+      match: { nameIncludes: 'Pocket Universe' },
+      listenMs: POCKET_UNIVERSE_EIP6963_LISTEN_MS,
+    }),
+    createExtensionDetector({
+      kind: 'pocketDatasetCsn',
+      observeMs: POCKET_UNIVERSE_HTML_DATA_CSN_SNAPSHOT_OBSERVE_MS,
+    }),
   ],
 };
