+++ src/providers/ExtensionDetectionProvider/withTimeout.ts
@@ -1,0 +1,9 @@
+export const withTimeout = <T>(
+  promise: Promise<T>,
+  ms: number,
+  fallback: T,
+): Promise<T> =>
+  Promise.race([
+    promise,
+    new Promise<T>((resolve) => setTimeout(() => resolve(fallback), ms)),
+  ]);
